refactor: moves external plugins to subpackage.
