+++ Makefile
@@ -10,7 +10,7 @@ build-plugins:
 	mkdir -p bin
 	for p in swaggerui; do \
 		echo "Building plugin $$p"; \
-		go build -tags lambda.norpc -ldflags "$(LDFLAGS)" -o ./bin/plugin-$$p ./external/$$p/impl; \
+		go build -tags lambda.norpc -ldflags "$(LDFLAGS)" -o ./bin/plugin-$$p ./external/plugins/$$p/impl; \
 	done
 
 .PHONY: fmt
