+++ external/plugins/swaggerui/impl/www/swagger-ui-es-bundle-core.js
