+++ external/plugins/swaggerui/impl/www/swagger-ui-standalone-preset.js
